Update fmt, googletest and jsoncpp submodules Update to C++23 Fix compile errors form breaking api changes

+++ src/Bindings/PluginLua.cpp
@@ -1148,7 +1148,7 @@ const char * cPluginLua::GetHookFnName(int a_HookType)
 			break;
 		}
 	}  // switch (a_Hook)
-	LOGWARNING("Requested name of an unknown hook type function: %d (max is %d)", a_HookType, cPluginManager::HOOK_MAX);
+	LOGWARNING("Requested name of an unknown hook type function: %d (max is %d)", static_cast<UInt32>(a_HookType), static_cast<UInt32>(cPluginManager::HOOK_MAX));
 	ASSERT(!"Unknown hook requested!");
 	return nullptr;
 }
